feat: implement Dashboard components including background, header, recent projects, and stats cards
