+++ frontend/src/components/dashboard/DashboardBackground.tsx
@@ -1,0 +1,74 @@
+import { Box } from '@mui/material';
+import React from 'react';
+
+export const DashboardBackground: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  return (
+    <>
+      <style>
+        {`
+          @keyframes float {
+            0%, 100% { transform: translateY(0) rotate(0deg); }
+            50% { transform: translateY(-20px) rotate(180deg); }
+          }
+          @keyframes floatSlow {
+            0%, 100% { transform: translateY(0) rotate(0deg); }
+            50% { transform: translateY(-15px) rotate(-180deg); }
+          }
+          @keyframes floatMedium {
+            0%, 100% { transform: translateY(0) rotate(0deg); }
+            50% { transform: translateY(-25px) rotate(90deg); }
+          }
+        `}
+      </style>
+      <Box
+        sx={{
+          minHeight: '100vh', // Full viewport height since navbar is now fixed
+          position: 'relative',
+          overflow: 'hidden',
+        }}
+      >
+        {/* Background Circles */}
+        <Box
+          sx={{
+            position: 'absolute',
+            top: '-20%',
+            right: '-10%',
+            width: '300px',
+            height: '300px',
+            borderRadius: '50%',
+            background: 'rgba(255, 255, 255, 0.1)',
+            animation: 'float 6s ease-in-out infinite',
+          }}
+        />
+        <Box
+          sx={{
+            position: 'absolute',
+            bottom: '-15%',
+            left: '-5%',
+            width: '200px',
+            height: '200px',
+            borderRadius: '50%',
+            background: 'rgba(255, 255, 255, 0.08)',
+            animation: 'floatSlow 8s ease-in-out infinite',
+          }}
+        />
+        <Box
+          sx={{
+            position: 'absolute',
+            top: '30%',
+            left: '80%',
+            width: '150px',
+            height: '150px',
+            borderRadius: '50%',
+            background: 'rgba(255, 255, 255, 0.06)',
+            animation: 'floatMedium 7s ease-in-out infinite',
+          }}
+        />
+
+        <Box sx={{ position: 'relative', zIndex: 1 }}>{children}</Box>
+      </Box>
+    </>
+  );
+};
