+++ frontend/src/components/dashboard/index.ts
@@ -1,0 +1,4 @@
+export { DashboardBackground } from './DashboardBackground';
+export { DashboardHeader } from './DashboardHeader';
+export { RecentProjects } from './RecentProjects';
+export { StatsCards } from './StatsCards';
